requirements.txt
streamlit
python-dotenv
litellm
crewai
crewai-tools